GP-15361 Improve error handling during import
This adds more graceful error handling for scenarios where the
received data set is empty, or the parsing logic makes it look like
it is empty. Instead of replacing the table data with an empty
data set, this will leave the existing data as-is and raise an
error instead.

+++ CRM/Postcodeat/ImportStatistikAustria.php
@@ -94,10 +94,15 @@ class CRM_Postcodeat_ImportStatistikAustria {
     return $newOrtnam;
   }
 
-  public function copy() {
+  public function copy($discardEmpty = TRUE) {
+    $check_sql = "SELECT COUNT(*) FROM `civicrm_statistikaustria_import`";
+    if ($discardEmpty && CRM_Core_DAO::singleValueQuery($check_sql) == 0) {
+      return FALSE;
+    }
     CRM_Core_DAO::executeQuery("TRUNCATE `civicrm_postcodeat`;");
     CRM_Core_DAO::executeQuery("INSERT INTO `civicrm_postcodeat` SELECT * FROM `civicrm_statistikaustria_import`");
     CRM_Core_DAO::executeQuery("TRUNCATE `civicrm_statistikaustria_import`;");
+    return TRUE;
   }
 
   /**
